fix(models/item): change decimal to float

+++ back-end/models/item.js
@@ -27,7 +27,7 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: false,
       },
       price: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT(10, 2),
         allowNull: false,
       },
     },
